Remove nested includes
