+++ src/Core/SettingsEnums.h
@@ -5,6 +5,7 @@
 #include <Core/LoadBalancing.h>
 #include <Core/LogsLevel.h>
 #include <Core/MergeSelectorAlgorithm.h>
+#include <Core/MergeTreeSerializationEnums.h>
 #include <Core/ParallelReplicasMode.h>
 #include <Core/QueryLogElementType.h>
 #include <Core/SchemaInferenceMode.h>
@@ -457,55 +458,12 @@ enum class GeoToH3ArgumentOrder : uint8_t
 
 DECLARE_SETTING_ENUM(GeoToH3ArgumentOrder)
 
-enum class MergeTreeSerializationInfoVersion : uint8_t
-{
-    BASIC = 0,
-    WITH_TYPES = 1,
-};
 
 DECLARE_SETTING_ENUM(MergeTreeSerializationInfoVersion)
-
-enum class MergeTreeStringSerializationVersion : uint8_t
-{
-    SINGLE_STREAM = 0,
-    WITH_SIZE_STREAM = 1,
-};
-
 DECLARE_SETTING_ENUM(MergeTreeStringSerializationVersion)
-
-enum class MergeTreeNullableSerializationVersion : uint8_t
-{
-    BASIC = 0,
-    ALLOW_SPARSE = 1,
-};
-
 DECLARE_SETTING_ENUM(MergeTreeNullableSerializationVersion)
-
-enum class MergeTreeObjectSerializationVersion : uint8_t
-{
-    V1,
-    V2,
-    V3,
-};
-
 DECLARE_SETTING_ENUM(MergeTreeObjectSerializationVersion)
-
-enum class MergeTreeObjectSharedDataSerializationVersion : uint8_t
-{
-    MAP,
-    MAP_WITH_BUCKETS,
-    ADVANCED,
-};
-
 DECLARE_SETTING_ENUM(MergeTreeObjectSharedDataSerializationVersion)
-
-enum class MergeTreeDynamicSerializationVersion : uint8_t
-{
-    V1,
-    V2,
-    V3,
-};
-
 DECLARE_SETTING_ENUM(MergeTreeDynamicSerializationVersion)
 
 enum class SearchOrphanedPartsDisks : uint8_t
