+++ src/DataTypes/Serializations/SerializationObject.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <Columns/ColumnObject.h>
+#include <Core/MergeTreeSerializationEnums.h>
 #include <DataTypes/Serializations/SerializationObjectSharedData.h>
 #include <Common/re2.h>
