+++ src/DataTypes/Serializations/SerializationObjectSharedData.h
@@ -5,6 +5,8 @@
 #include <DataTypes/Serializations/SimpleTextSerialization.h>
 #include <boost/algorithm/string/join.hpp>
 
+#include <map>
+
 namespace DB
 {
