+++ src/DataTypes/Serializations/SerializationString.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <Core/MergeTreeSerializationEnums.h>
 #include <DataTypes/Serializations/ISerialization.h>
 
 namespace DB
