+++ src/Formats/NativeReader.h
@@ -7,6 +7,8 @@
 #include <Core/Block.h>
 #include <Core/BlockMissingValues.h>
 
+#include <map>
+
 namespace DB
 {
